backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
python-multipart==0.0.9
pandas==2.2.2
openpyxl==3.1.5
httpx==0.27.0
python-dotenv==1.0.1
scipy==1.13.1
